Data Import Debug
Use for DEBUG MODE only.

Update the following URL with the path to your installation directory
and the domain you wish to capture data from.
http://localhost/organic-search-analytics/data-capture-run.php?type=goog
leSearchAnalytics&domain=yourwebsite.com&date=2015-12-30

WARNING - this will import data for the date selected and thus will
have skew your data for that day.  This should be performed on a test
database/installation else be prepared to delete the false data that is
captured for that day.