Delete Data Functionality
A page that allows the user to delete data from Google for dates where
the data can be recaptured.  This is to use only if you feel there is a
discrepancy with the data you have from Google.  This provides a quick
easy way for users affected by issue #10 to be able to quickly delete
data that they can then re-capture from Google.  Use at your own risk as
your are deleting data from your own database.