2.4.0
Capture country from Google and allow filtering of reports by a specific
country.  Fixes #11

Notice - you'll need to run an update script from the home page for this
version to run smoothly as it needs to update the database.

The initial report settings page is a bit "odd" when selecting a country
as it will display ALL countries available from all domains, queries,
date ranges, etc.  For that reason, it'd be possible to generate a
report that pulls back no results.  However, once viewing a report, we
have some information to base the country validation off of and thus the
list is filtered down at that point.  So it may be best practice to
generate a report with no country selected and then add the country in a
second step.  Though if you are looking at the primary country that
drives traffic to your site this may be a moot point.