Correct $GLOBALS directory structure variables
Revert code that was proposed in #31, ported to #38, and mistakenly deployed via 9d72b4d that attempts to work with an Apache alias environment.  The proposed code modified the $GLOBALS['basedirWebServer'] and $GLOBALS['appInstallDir'] variables, which resulted in unexpected values in $GLOBALS['scriptName'].  The final result of this code change is that under certain Apache configurations it was not possible to reliably use $GLOBALS['scriptName'] to detect what page was being rendered and thus conditional situations were failing.  The example brought up in #51 is that a javascript file was not being loaded on the reports.php page.